% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cells_with_pmp.R
\name{cells_with_pmp}
\alias{cells_with_pmp}
\title{produce a table with cells exhibiting given proteins on plasma
membrane according to CL}
\usage{
cells_with_pmp(curies)
}
\arguments{
\item{curies}{a character vector in format "PR:nnnnnnnnn"}
}
\value{
a data.frame with columns cl, celltype, pr, protein
}
\description{
produce a table with cells exhibiting given proteins on plasma
membrane according to CL
}
\examples{
cells_with_pmp(c("PR:000002064", "PR:000001874"))
}
