% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{cn2tag}
\alias{cn2tag}
\title{a named vector with mapping from cell type phrase to CURIE for CL.owl}
\format{
names character vector
}
\usage{
data(cn2tag)
}
\description{
a named vector with mapping from cell type phrase to CURIE for CL.owl
}
\note{
creation is detailed in unexported function makeBundledMaps
}
\examples{
data("cn2tag", package = "ontoProc2")
cn2tag["Kupffer cell"]
}
\keyword{datasets}
