% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SemsqlConn-S7.R
\name{count_descendants}
\alias{count_descendants}
\title{Count the number of descendants of a term}
\usage{
count_descendants(x, ...)
}
\arguments{
\item{x}{A \code{SemsqlConn} object.}

\item{term_id}{character(1) CURIE.}

\item{predicate}{character(1) predicate to traverse
(default \code{"rdfs:subClassOf"}).}
}
\value{
integer(1).
}
\description{
Count the number of descendants of a term
}
\examples{
goref <- semsql_connect(ontology = "go")
count_descendants(goref, "GO:0006915") # all apoptosis subtypes
disconnect(goref)
}
