% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SemsqlConn-S7.R
\name{describe_table}
\alias{describe_table}
\title{Describe the columns of a table in a SemsqlConn database}
\usage{
describe_table(x, ...)
}
\arguments{
\item{x}{A \code{SemsqlConn} object.}

\item{table_name}{character(1) name of the table.}
}
\value{
data.frame with PRAGMA table_info output (columns: cid, name, type,
  notnull, dflt_value, pk).
}
\description{
Describe the columns of a table in a SemsqlConn database
}
\examples{
goref <- semsql_connect(ontology = "go")
describe_table(goref, "rdfs_label_statement")
disconnect(goref)
}
