% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SemsqlConn-S7.R
\name{find_by_restriction}
\alias{find_by_restriction}
\title{Find terms that have a given OWL someValuesFrom restriction}
\usage{
find_by_restriction(x, ...)
}
\arguments{
\item{x}{A \code{SemsqlConn} object.}

\item{property}{character(1) property CURIE
(e.g. \code{"BFO:0000050"} for part-of).}

\item{filler}{character(1) filler class CURIE.}

\item{include_filler_descendants}{logical(1) if TRUE also match subclasses
of \code{filler} (default \code{FALSE}).}
}
\value{
data.frame with columns \code{id} and \code{label}.
}
\description{
Find terms that have a given OWL someValuesFrom restriction
}
\examples{
goref <- semsql_connect(ontology = "go")
# cellular components that are part_of nucleus (GO:0005634)
find_by_restriction(goref, "BFO:0000050", "GO:0005634")
disconnect(goref)
}
