% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SemsqlConn-S7.R
\name{get_definition}
\alias{get_definition}
\title{Get the text definition for a term}
\usage{
get_definition(x, ...)
}
\arguments{
\item{x}{A \code{SemsqlConn} object.}

\item{term_id}{character(1) CURIE.}
}
\value{
character(1) definition text, or \code{NA_character_} if not found.
}
\description{
Get the text definition for a term
}
\examples{
goref <- semsql_connect(ontology = "go")
get_definition(goref, "GO:0006915")
disconnect(goref)
}
