% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SemsqlConn-S7.R
\name{get_direct_superclasses}
\alias{get_direct_superclasses}
\title{Get direct superclasses of a term}
\usage{
get_direct_superclasses(x, ...)
}
\arguments{
\item{x}{A \code{SemsqlConn} object.}

\item{term_id}{character(1) CURIE.}
}
\value{
data.frame with columns \code{id} and \code{label}, ordered by
  label.
}
\description{
Get direct superclasses of a term
}
\examples{
goref <- semsql_connect(ontology = "go")
get_direct_superclasses(goref, "GO:0006915")
disconnect(goref)
}
