% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SemsqlConn-S7.R
\name{get_label}
\alias{get_label}
\title{Get the rdfs:label for a term}
\usage{
get_label(x, ...)
}
\arguments{
\item{x}{A \code{SemsqlConn} object.}

\item{term_id}{character(1) CURIE, e.g. \code{"CL:0000540"}.}
}
\value{
character(1) label, or \code{NA_character_} if not found.
}
\description{
Get the rdfs:label for a term
}
