% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SemsqlConn-S7.R
\name{get_prefix}
\alias{get_prefix}
\title{Retrieve the ontology prefix from a SemsqlConn}
\usage{
get_prefix(x, ...)
}
\arguments{
\item{x}{A \code{SemsqlConn} object.}
}
\value{
character(1) the primary ontology prefix (e.g. \code{"CL"}).
}
\description{
Retrieve the ontology prefix from a SemsqlConn
}
