% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_present_pmp.R
\name{get_present_pmp}
\alias{get_present_pmp}
\title{produce a table with list of proteins from protein ontology
identified as present on cell membranes for input cell type CURIEs}
\usage{
get_present_pmp(curies)
}
\arguments{
\item{curies}{a character vector in format "CL:nnnnnnn"}
}
\value{
a data.frame with columns cl, celltype, pr, protein
}
\description{
produce a table with list of proteins from protein ontology
identified as present on cell membranes for input cell type CURIEs
}
\examples{
get_present_pmp(c("CL:0000091", "CL:0000926"))
}
