% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SemsqlConn-S7.R
\name{get_restrictions}
\alias{get_restrictions}
\title{Get OWL someValuesFrom restrictions for a term}
\usage{
get_restrictions(x, ...)
}
\arguments{
\item{x}{A \code{SemsqlConn} object.}

\item{term_id}{character(1) CURIE.}
}
\value{
data.frame with columns \code{restriction_id}, \code{property},
  \code{property_label}, \code{filler}, \code{filler_label}.
}
\description{
Get OWL someValuesFrom restrictions for a term
}
\examples{
goref <- semsql_connect(ontology = "go")
# mitochondrion (GO:0005739) has part-of restrictions to cell
get_restrictions(goref, "GO:0005739")
disconnect(goref)
}
