% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SemsqlConn-S7.R
\name{get_term_info}
\alias{get_term_info}
\title{Retrieve a summary of information about a term}
\usage{
get_term_info(x, ...)
}
\arguments{
\item{x}{A \code{SemsqlConn} object.}

\item{term_id}{character(1) CURIE.}
}
\value{
list with elements \code{id}, \code{label}, \code{definition},
  \code{synonyms}, \code{superclasses}, \code{subclasses}.
}
\description{
Retrieve a summary of information about a term
}
\examples{
goref <- semsql_connect(ontology = "go")
info <- get_term_info(goref, "GO:0006915")
info$label
info$superclasses
disconnect(goref)
}
