% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graphNEL.R
\name{improveNodes}
\alias{improveNodes}
\title{inject linefeeds for node names for graph, with textual
annotation from ontology}
\usage{
improveNodes(g, ont)
}
\arguments{
\item{g}{graphNEL instance}

\item{ont}{instance of ontology from ontologyIndex}
}
\value{
graphNEL instance
}
\description{
inject linefeeds for node names for graph, with textual
annotation from ontology
}
\examples{
requireNamespace("Rgraphviz")
requireNamespace("graph")
clcon <- retrieve_semsql_conn("cl")
cl <- semsql_to_oi(clcon)
cl3k <- c(
  "CL:0000492", "CL:0001054", "CL:0000236", "CL:0000625",
  "CL:0000576", "CL:0000623", "CL:0000451", "CL:0000556"
)
p3k <- ontologyPlot::onto_plot(cl, cl3k)
gnel <- make_graphNEL_from_ontology_plot(p3k)
head(graph::nodes(gnel)) # before improving
gnel <- improveNodes(gnel, cl)
head(graph::nodes(gnel))
}
