% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggraph_viz.R
\name{label_igraph_nodes}
\alias{label_igraph_nodes}
\title{Add term labels as a vertex attribute to an ontology igraph}
\usage{
label_igraph_nodes(g, ont, sep = "\\n")
}
\arguments{
\item{g}{igraph directed graph whose vertex names are ontology term IDs}

\item{ont}{ontology_index instance (from ontologyIndex)}

\item{sep}{character(1) separator between term name and ID, defaults to "\n"}
}
\value{
igraph with a "label" vertex attribute
}
\description{
Add term labels as a vertex attribute to an ontology igraph
}
\examples{
requireNamespace("igraph")
clcon <- retrieve_semsql_conn("cl")
cl <- semsql_to_oi(clcon)
cl3k <- c(
  "CL:0000492", "CL:0001054", "CL:0000236", "CL:0000625",
  "CL:0000576", "CL:0000623", "CL:0000451", "CL:0000556"
)
p3k <- ontologyPlot::onto_plot(cl, cl3k)
ig <- make_igraph_from_ontology_plot(p3k)
ig <- label_igraph_nodes(ig, cl)
igraph::V(ig)$label
}
