% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SemsqlConn-S7.R
\name{list_tables}
\alias{list_tables}
\title{List tables in a SemsqlConn database}
\usage{
list_tables(x, ...)
}
\arguments{
\item{x}{A \code{SemsqlConn} object.}
}
\value{
character vector of table names.
}
\description{
List tables in a SemsqlConn database
}
