% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggraph_viz.R
\name{make_igraph_from_ontology_plot}
\alias{make_igraph_from_ontology_plot}
\title{Convert an ontology_plot instance to an igraph object}
\usage{
make_igraph_from_ontology_plot(x)
}
\arguments{
\item{x}{instance of S3 class ontology_plot (from ontologyPlot::onto_plot)}
}
\value{
an igraph directed graph
}
\description{
Convert an ontology_plot instance to an igraph object
}
\examples{
requireNamespace("igraph")
clcon <- retrieve_semsql_conn("cl")
cl <- semsql_to_oi(clcon)
cl3k <- c(
  "CL:0000492", "CL:0001054", "CL:0000236", "CL:0000625",
  "CL:0000576", "CL:0000623", "CL:0000451", "CL:0000556"
)
p3k <- ontologyPlot::onto_plot(cl, cl3k)
ig <- make_igraph_from_ontology_plot(p3k)
ig
}
