% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ontoProc2_package.R
\docType{package}
\name{ontoProc2-package}
\alias{ontoProc2}
\alias{ontoProc2-package}
\title{ontoProc2: Ontology Facilities Based on INCAtools Semantic SQL}
\description{
This package provides ontology facilities based on INCAtools Semantic SQL. Tooling is provided to retrieve and cache SQLite databases representing curated ontologies. ontologyIndex ontology_index instances can be produced. Anthropic's Claude was used in the development of S7 classes and methods and in the production of documentation.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/vjcitn/ontoProc2}
  \item Report bugs at \url{https://github.com/vjcitn/ontoProc2/issues}
}

}
\author{
\strong{Maintainer}: Vincent Carey \email{stvjc@channing.harvard.edu} (\href{https://orcid.org/0000-0003-4046-0063}{ORCID})

Authors:
\itemize{
  \item Vincent Carey \email{stvjc@channing.harvard.edu} (\href{https://orcid.org/0000-0003-4046-0063}{ORCID})
}

Other contributors:
\itemize{
  \item Anthropic-Claude Sonnet4.6 [engineer]
  \item NIH NHGRI 2U24HG004059-17 [funder]
}

}
\keyword{internal}
