% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SemsqlConn-S7.R
\docType{data}
\name{PREDICATES}
\alias{PREDICATES}
\title{Standard predicate CURIEs used in OBO ontologies}
\format{
A named list with elements \code{subclass_of}, \code{part_of},
  \code{has_part}, \code{develops_from}, \code{located_in},
  \code{has_characteristic}.
}
\usage{
PREDICATES
}
\description{
A named list of commonly used predicate CURIEs in OBO-format ontologies,
for use with \code{\link{get_ancestors}}, \code{\link{get_descendants}},
and related functions.
}
\examples{
# Connect to the Cell Ontology and retrieve ancestors using part-of edges
conn <- semsql_connect(ontology = "cl")
# neuron (CL:0000540) ancestors via is-a and part-of
anc <- get_ancestors(conn, "CL:0000540",
  predicates = c(PREDICATES$subclass_of, PREDICATES$part_of))
head(anc)
disconnect(conn)
}
\keyword{datasets}
