% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SemsqlConn-S7.R
\name{run_query}
\alias{run_query}
\title{Run an arbitrary SQL query against a SemsqlConn database}
\usage{
run_query(x, ...)
}
\arguments{
\item{x}{A \code{SemsqlConn} object.}

\item{sql}{character(1) SQL query string.}
}
\value{
data.frame with query results.
}
\description{
Run an arbitrary SQL query against a SemsqlConn database
}
