% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SemSQL-S4.R
\name{select,SemSQL-method}
\alias{select,SemSQL-method}
\title{emulate AnnotationDbi}
\usage{
\S4method{select}{SemSQL}(x, keys, columns, keytype = "GOID", ...)
}
\description{
emulate AnnotationDbi
}
\examples{
gg = retrieve_semsql_conn("go")
ngo = SemSQL(gg, "GO")
ngo
validObject(ngo)
select(ngo, keys=c("GO:0018942", "GO:0000003"), columns=c("DEFINITION", "TERM", "ONTOLOGY"))
}
