% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SemSQL-S4.R
\name{select}
\alias{select}
\title{emulate the AnnotationDbi select method}
\usage{
select(x, keys, columns, keytype = "GOID", ...)
}
\arguments{
\item{x}{instance of SemSQL}

\item{keys}{vector of elements of the appropriate type}

\item{columns}{vector of desired output columns}

\item{keytype}{character(1) defaults to 'GOID'}

\item{\dots}{not used}
}
\description{
emulate the AnnotationDbi select method
}
