% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SemSQL-S4.R
\docType{class}
\name{SemSQL-class}
\alias{SemSQL-class}
\title{define a wrapper for the SQLite connection}
\description{
define a wrapper for the SQLite connection
}
\section{Slots}{

\describe{
\item{\code{conn}}{an instance of SQLiteConnection}

\item{\code{resource}}{character(1) a descriptive string}

\item{\code{nstats}}{numeric(1) holder for count of number of statements}
}}

