% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SemSQL-S4.R
\name{SemSQL}
\alias{SemSQL}
\title{constructor for SemSQL instance}
\usage{
SemSQL(conn, resource)
}
\arguments{
\item{conn}{SQLiteConnection}

\item{resource}{character tag}
}
\description{
constructor for SemSQL instance

check aspects of SemSQL object
}
