% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/semsql_url.R
\name{semsql_url}
\alias{semsql_url}
\title{produce INCAtools distribution URL}
\usage{
semsql_url(ontology = "efo")
}
\arguments{
\item{ontology}{short string that is the prefix to .db.gz in the bbop-sqlite collection}
}
\description{
produce INCAtools distribution URL
}
\examples{
semsql_url("cl")
}
