% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SemSQL-S4.R
\name{show,SemSQL-method}
\alias{show,SemSQL-method}
\title{present information about a SemSQL object}
\usage{
\S4method{show}{SemSQL}(object)
}
\arguments{
\item{object}{instance of SemSQL}
}
\description{
present information about a SemSQL object
}
