% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bbop_sqlite_db_gz.R
\name{bbop_sqlite_db_gz}
\alias{bbop_sqlite_db_gz}
\title{retrieve and cache all filenames of Semantic SQL ontologies available, checking for updated
content relative to cache via ETag}
\usage{
bbop_sqlite_db_gz(
  url = "https://s3.amazonaws.com/bbop-sqlite",
  bfc = BiocFileCache::BiocFileCache(),
  rname = "bbop_sqlite_listing",
  timeout_connect = 5,
  timeout_total = 15
)
}
\arguments{
\item{url}{source address}

\item{bfc}{an instance of BiocFileCache}

\item{rname}{a string to use to identify the cached listing}

\item{timeout_connect}{passed to httr2::req_options connecttimeout parameter}

\item{timeout_total}{passed to httr2::req_options seconds parameter}
}
\description{
retrieve and cache all filenames of Semantic SQL ontologies available, checking for updated
content relative to cache via ETag
}
\note{
There is no real etag discipline for the metadata, so full metadata
content is extracted on each call, digested, and compared to digest in cache.
Cache is updated when there is a discrepancy, except there is a guard against
rewriting cache with corrupted or zero-length data at endpoint.  Note this
code was negotiated at length with claude.ai Sonnet 4.6.
}
\examples{
gzs = bbop_sqlite_db_gz()
head(gzs)
length(gzs)
}
