% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SemsqlConn-S7.R
\name{disconnect}
\alias{disconnect}
\title{Disconnect a SemsqlConn from its database}
\usage{
disconnect(x, ...)
}
\arguments{
\item{x}{A \code{SemsqlConn} object.}

\item{quiet}{logical(1) if TRUE suppresses the disconnection message
(default FALSE).}
}
\value{
The \code{SemsqlConn} object invisibly.
}
\description{
Disconnect a SemsqlConn from its database
}
\examples{
goref <- semsql_connect(ontology = "go")
disconnect(goref)
}
