% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SemsqlConn-S7.R
\name{find_intersection}
\alias{find_intersection}
\title{Find terms that are descendants of a superclass and have a given restriction}
\usage{
find_intersection(x, ...)
}
\arguments{
\item{x}{A \code{SemsqlConn} object.}

\item{superclass_id}{character(1) CURIE of the superclass.}

\item{relation_property}{character(1) property CURIE for the restriction.}

\item{related_to_id}{character(1) filler CURIE for the restriction.}
}
\value{
data.frame with columns \code{id} and \code{label}.
}
\description{
Find terms that are descendants of a superclass and have a given restriction
}
