% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SemsqlConn-S7.R
\name{get_direct_edges}
\alias{get_direct_edges}
\title{Get direct edges in the ontology graph for a term}
\usage{
get_direct_edges(x, ...)
}
\arguments{
\item{x}{A \code{SemsqlConn} object.}

\item{term_id}{character(1) CURIE.}

\item{direction}{character(1) one of \code{"outgoing"}, \code{"incoming"},
\code{"both"}.}
}
\value{
data.frame with columns \code{subject}, \code{subject_label},
  \code{predicate}, \code{predicate_label}, \code{object},
  \code{object_label}.
}
\description{
Get direct edges in the ontology graph for a term
}
\examples{
goref <- semsql_connect(ontology = "go")
get_direct_edges(goref, "GO:0006915")
get_direct_edges(goref, "GO:0006915", direction = "both")
disconnect(goref)
}
