% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SemsqlConn-S7.R
\name{get_direct_subclasses}
\alias{get_direct_subclasses}
\title{Get direct subclasses of a term}
\usage{
get_direct_subclasses(x, ...)
}
\arguments{
\item{x}{A \code{SemsqlConn} object.}

\item{term_id}{character(1) CURIE.}
}
\value{
data.frame with columns \code{id} and \code{label}, ordered by
  label.
}
\description{
Get direct subclasses of a term
}
