% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SemsqlConn-S7.R
\name{get_synonyms}
\alias{get_synonyms}
\title{Get synonyms for a term}
\usage{
get_synonyms(x, ...)
}
\arguments{
\item{x}{A \code{SemsqlConn} object.}

\item{term_id}{character(1) CURIE.}

\item{type}{character(1) synonym scope: one of \code{"all"},
\code{"exact"}, \code{"broad"}, \code{"narrow"}, \code{"related"}.}
}
\value{
data.frame with columns \code{subject}, \code{predicate},
  \code{synonym}.
}
\description{
Get synonyms for a term
}
\examples{
goref <- semsql_connect(ontology = "go")
get_synonyms(goref, "GO:0006915")
get_synonyms(goref, "GO:0006915", type = "exact")
disconnect(goref)
}
