% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SemsqlConn-S7.R
\name{is_connected}
\alias{is_connected}
\title{Test whether a SemsqlConn has a valid open connection}
\usage{
is_connected(x, ...)
}
\arguments{
\item{x}{A \code{SemsqlConn} object.}
}
\value{
logical(1).
}
\description{
Test whether a SemsqlConn has a valid open connection
}
