% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ncit_map}
\alias{ncit_map}
\title{a named vector with values rdfs labels in NCI thesaurus, and names the corresponding formal ontology tags}
\format{
named character vector
}
\usage{
data(ncit_map)
}
\description{
a named vector with values rdfs labels in NCI thesaurus, and names the corresponding formal ontology tags
}
\examples{
data("ncit_map", package="ontoProc2")
ncit_map["EFO:1000899"]
}
\keyword{datasets}
