% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{tag2cn}
\alias{tag2cn}
\title{a named vector with mapping from CURIE to cell type phrase for CL.owl of 2025-12-17}
\format{
names character vector
}
\usage{
data(tag2cn)
}
\description{
a named vector with mapping from CURIE to cell type phrase for CL.owl of 2025-12-17
}
\examples{
data("tag2cn", package = "ontoProc2")
tag2cn[c("CL:0000000", "CL:0000006")]
}
\keyword{datasets}
